requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
tzlocal==4.2
whitenoise==6.2.0
